+++ src/LinkDotNet.StringBuilder/ValueStringBuilder.AppendJoin.cs
@@ -17,6 +17,24 @@ public ref partial struct ValueStringBuilder
     /// <summary>
     /// Concatenates and appends all values with the given separator between each entry at the end of the string.
     /// </summary>
+    /// <param name="separator">String used as separator between the entries.</param>
+    /// <param name="values">Enumerable of strings to be concatenated.</param>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public void AppendJoin(ReadOnlySpan<char> separator, scoped ReadOnlySpan<string?> values)
+        => AppendJoinInternalString(separator, values);
+
+    /// <summary>
+    /// Concatenates and appends all values with the given separator between each entry at the end of the string.
+    /// </summary>
+    /// <param name="separator">Character used as separator between the entries.</param>
+    /// <param name="values">Enumerable of strings to be concatenated.</param>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public void AppendJoin(char separator, scoped ReadOnlySpan<string?> values)
+        => AppendJoinInternalChar(separator, values);
+
+    /// <summary>
+    /// Concatenates and appends all values with the given separator between each entry at the end of the string.
+    /// </summary>
     /// <param name="separator">Character used as separator between the entries.</param>
     /// <param name="values">Enumerable of strings to be concatenated.</param>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
@@ -39,7 +57,17 @@ public ref partial struct ValueStringBuilder
     /// <param name="values">Enumerable to be concatenated.</param>
     /// <typeparam name="T">Type of the given enumerable.</typeparam>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public void AppendJoin<T>(ReadOnlySpan<char> separator, IEnumerable<T> values)
+    public void AppendJoin<T>(scoped ReadOnlySpan<char> separator, IEnumerable<T> values)
+        => AppendJoinInternalString(separator, values);
+
+    /// <summary>
+    /// Concatenates and appends all values with the given separator between each entry at the end of the string.
+    /// </summary>
+    /// <param name="separator">String used as separator between the entries.</param>
+    /// <param name="values">Enumerable to be concatenated.</param>
+    /// <typeparam name="T">Type of the given enumerable.</typeparam>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public void AppendJoin<T>(scoped ReadOnlySpan<char> separator, ReadOnlySpan<T> values)
         => AppendJoinInternalString(separator, values);
 
     /// <summary>
@@ -55,6 +83,16 @@ public ref partial struct ValueStringBuilder
     /// <summary>
     /// Concatenates and appends all values with the given separator between each entry at the end of the string.
     /// </summary>
+    /// <param name="separator">Character used as separator between the entries.</param>
+    /// <param name="values">Enumerable to be concatenated.</param>
+    /// <typeparam name="T">Type of the given enumerable.</typeparam>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public void AppendJoin<T>(char separator, scoped ReadOnlySpan<T> values)
+        => AppendJoinInternalChar(separator, values);
+
+    /// <summary>
+    /// Concatenates and appends all values with the given separator between each entry at the end of the string.
+    /// </summary>
     /// <param name="separator">Rune used as separator between the entries.</param>
     /// <param name="values">Enumerable to be concatenated.</param>
     /// <typeparam name="T">Type of the given enumerable.</typeparam>
@@ -63,7 +101,7 @@ public ref partial struct ValueStringBuilder
         => AppendJoinInternalRune(separator, values);
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private void AppendJoinInternalString<T>(ReadOnlySpan<char> separator, IEnumerable<T> values)
+    private void AppendJoinInternalString<T>(scoped ReadOnlySpan<char> separator, IEnumerable<T> values)
     {
         ArgumentNullException.ThrowIfNull(values);
 
@@ -82,6 +120,23 @@ public ref partial struct ValueStringBuilder
             Append(separator);
             current = enumerator.Current;
             AppendInternal(current);
+        }
+    }
+
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    private void AppendJoinInternalString<T>(scoped ReadOnlySpan<char> separator, scoped ReadOnlySpan<T> values)
+    {
+        if (values.Length == 0)
+        {
+            return;
+        }
+
+        AppendInternal(values[0]);
+
+        for (var i = 1; i < values.Length; i++)
+        {
+            Append(separator);
+            AppendInternal(values[i]);
         }
     }
 
@@ -105,6 +160,23 @@ public ref partial struct ValueStringBuilder
             AppendInternal(separator);
             current = enumerator.Current;
             AppendInternal(current);
+        }
+    }
+
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    private void AppendJoinInternalChar<T>(char separator, scoped ReadOnlySpan<T> values)
+    {
+        if (values.Length == 0)
+        {
+            return;
+        }
+
+        AppendInternal(values[0]);
+
+        for (var i = 1; i < values.Length; i++)
+        {
+            Append(separator);
+            AppendInternal(values[i]);
         }
     }